% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreBerglas.R
\name{scoreBerglas}
\alias{scoreBerglas}
\title{A function that scores Berglas et al.'s Sexual Relationship Rights (SRR) measure}
\usage{
scoreBerglas(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "Q2_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Subscale and total average scores
}
\examples{
dat = scoreBerglas(dat, firstq = "AA_SSR_1")
}
