% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreBuunk.R
\name{scoreBuunk}
\alias{scoreBuunk}
\title{A function that scores Buunk's Types of Jealousy Scale}
\usage{
scoreBuunk(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "APCBSSI_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates Reactive, Preventative, and Anxious subscale scores
}
\examples{
dat = scoreBuunk(dat, firstq = "APCBSSI_1")
}
