% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreCamilleri.R
\name{scoreCamilleri}
\alias{scoreCamilleri}
\title{A function that scores Camilleri's Tactics to Obtain Sex Scale}
\usage{
scoreCamilleri(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "TOSS_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates coerce, coax, and total scores
}
\examples{
dat = scoreCamilleri(dat, firstq = "TOSS_1")
}
