% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreCataniaDSC.R
\name{scoreCataniaDSC}
\alias{scoreCataniaDSC}
\title{A function that scores Catania's Dyadic Sex Communication Scale}
\usage{
scoreCataniaDSC(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "DSCS_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates total score
}
\examples{
dat = scoreCataniaDSC(dat, firstq = "DSCS_1")
}
