% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreCataniaDySexReg.R
\name{scoreCataniaDySexReg}
\alias{scoreCataniaDySexReg}
\title{A function that scores Catania's Dyadic Sex Regulation Scale}
\usage{
scoreCataniaDySexReg(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "DSRS_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates scale score
}
\examples{
dat = scoreCataniaDySexReg(dat, firstq = "DSRS_1")
}
