% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreDharma.R
\name{scoreDharma}
\alias{scoreDharma}
\title{A function that scores Dharma et al's Trans-Specific Body Image Worries Scale}
\usage{
scoreDharma(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "TSSBIW_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates total score
}
\examples{
dat = scoreDharma(dat, firstq = "TSSBIW_1")
}
