% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreDossett.R
\name{scoreDosset}
\alias{scoreDosset}
\title{A function that scores Dossett's Sexual Importance Scale}
\usage{
scoreDosset(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "SIS_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates scale score
}
\examples{
dat = scoreDosset(dat, firstq = "SIS_1")
}
