% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreEriksson.R
\name{scoreEriksson}
\alias{scoreEriksson}
\title{A function that scores Eriksson & Humphrey's Virginity Beliefs Scale}
\usage{
scoreEriksson(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "VBS_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates gift, stigma, and process subscale scores
}
\examples{
dat = scoreEriksson(dat, firstq = "VBS_1")
}
