% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreFaaborgAndersen.R
\name{scoreFaaborgAndersen}
\alias{scoreFaaborgAndersen}
\title{A function that scores Faaborg-Andersen et al.'s Index of Male Genital Image}
\usage{
scoreFaaborgAndersen(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "IMGI_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates five subscale scores and total score
}
\examples{
dat = scoreFaaborgAndersen(dat, firstq = "IMGI_1")
}
