% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreFisherATS.R
\name{scoreFisherATS}
\alias{scoreFisherATS}
\title{A function that scores Fisher's Attitude Towards Sexuality Scale}
\usage{
scoreFisherATS(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "ATSS_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates scale score
}
\examples{
dat = scoreFisherATS(dat, firstq = "ATSS_1")
}
