% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreGiambra.R
\name{scoreGiambra}
\alias{scoreGiambra}
\title{A function that scores Giambra & Singer's Sexual Daydreaming Scale}
\usage{
scoreGiambra(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "SDSIPI_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates scale score
}
\examples{
dat = scoreGiambra(dat, firstq = "SDSIPI_1")
}
