% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreGoetz.R
\name{scoreGoetz}
\alias{scoreGoetz}
\title{A function that scores Goetz & Shackleford's Sexual Coercion in Intimate Relationships Scale}
\usage{
scoreGoetz(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "SCIRS_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates three subscale and scale total scores
}
\examples{
dat = scoreGoetz(dat, firstq = "SCIRS_1")
}
