% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreGoldsmith.R
\name{scoreGoldsmith}
\alias{scoreGoldsmith}
\title{A function that scores Goldsmith and Byer's Sexual Idealization Scale}
\usage{
scoreGoldsmith(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "SIS_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates scale score
}
\examples{
dat = scoreGoldsmith(dat, firstq = "SIS_1")
}
