% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreGraham.R
\name{scoreGraham}
\alias{scoreGraham}
\title{A function that scores Graham et al.'s Sexual Excitation/Sexual Inhibition Inventory for
Women}
\usage{
scoreGraham(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "SAI_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Subscale average scores and higher order average scores
}
\examples{
dat = scoreGraham(dat, firstq = "SESIW1")
}
