% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreHendrick.R
\name{scoreHendrick}
\alias{scoreHendrick}
\title{A function that scores Hendrick & Hendrick's Brief Sexual Attitudes Scale}
\usage{
scoreHendrick(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "SAS_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates gift, stigma, and process subscale scores
}
\examples{
dat = scoreHendrick(dat, firstq = "SAS_1")
}
