% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreHoon.R
\name{scoreHoon}
\alias{scoreHoon}
\title{A function that scores Hoon et al.'s Sexual Arousability Inventory}
\usage{
scoreHoon(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "SAI_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Subscale sum scores
}
\examples{
dat = scoreHoon(dat, firstq = "SAI_1")
}
