% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreHumphreysFSRS.R
\name{scoreHumphreysFSRS}
\alias{scoreHumphreysFSRS}
\title{A function that scores Humphreys & Kennett's Female Sexual Resourcefulness Scale}
\usage{
scoreHumphreysFSRS(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "FSRS_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates total score
}
\examples{
dat = scoreHumphreysFSRS(dat, firstq = "FSRS_1")
}
