% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreHumphreysSCSR.R
\name{scoreHumphreysSCSR}
\alias{scoreHumphreysSCSR}
\title{A function that scores Humphreys's Sexual Consent Scale-Revised}
\usage{
scoreHumphreysSCSR(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "SCSR_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates five subscale scores
}
\examples{
dat = scoreHumphreysSCSR(dat, firstq = "SCSR_1")
}
