% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreJanssenSESI.R
\name{scoreJanssenSESI}
\alias{scoreJanssenSESI}
\title{A function that scores Janssen et al.'s Sexual Excitation/Sexual Inhibition Inventory}
\usage{
scoreJanssenSESI(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "SESIW1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Subscale average scores and higher order average scores
}
\examples{
dat = scoreJanssenSESI(dat, firstq = "SESIW1")
}
