% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreJanssenSESI_sf.R
\name{scoreJanssenSESI_SF}
\alias{scoreJanssenSESI_SF}
\title{A function that scores Janssen et al.'s Sexual Excitation/Sexual Inhibition Inventory (short-form)}
\usage{
scoreJanssenSESI_SF(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "sesi1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Subscale total scores
}
\examples{
dat = scoreJanssenSESI_SF(dat, firstq = "sesi1")
}
