% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreJozkowski.R
\name{scoreJozkowski}
\alias{scoreJozkowski}
\title{A function that scores Jozkowski's Internal and External Consent Scale}
\usage{
scoreJozkowski(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "ICS_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates subscale scores for internal subscales (external TBD)
}
\examples{
dat = scoreJozkowski(dat, firstq = "ICS_1")
}
