% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreKirby.R
\name{scoreKirby}
\alias{scoreKirby}
\title{A function that scores Kirby's Mathtech Questionnaire}
\usage{
scoreKirby(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "Q2_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculate's knowledge, attitude subscale scores.
}
\examples{
dat = scoreKirby(dat, firstq = "Q2_1")
}
