% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreLiss.R
\name{scoreLiss}
\alias{scoreLiss}
\title{A function that scores Liss et al's Enjoyment of Sexualization Score}
\usage{
scoreLiss(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "ESS_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates total score
}
\examples{
dat = scoreLiss(dat, firstq = "ESS_1")
}
