% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreLottesSSI.R
\name{scoreLottesSSI}
\alias{scoreLottesSSI}
\title{A function that scores Lottes et al.'s Sexual Socialization Instrument (SSI) measure}
\usage{
scoreLottesSSI(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "SSI_1")}
}
\value{
data frame with new vectors of subscale and overall scores
}
\description{
Subscale and total average scores
}
\examples{
dat = scoreLottesSSI(dat, firstq = "SSI_1")
}
