% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMarelichSDS.R
\name{scoreMarelichSDS}
\alias{scoreMarelichSDS}
\title{A function that scores Marelich et al.'s Sexual Deception Scale}
\usage{
scoreMarelichSDS(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "SDS_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates subscale and total scores
}
\examples{
dat = scoreMarelichSDS(dat, firstq = "SDS_1")
}
