% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMaxwell.R
\name{scoreMaxwell}
\alias{scoreMaxwell}
\title{A function that scores Maxwell et al.'s Implicit Theories of Sexuality Scale}
\usage{
scoreMaxwell(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "ITSS_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates growth and destiny subscale scores
}
\examples{
dat = scoreMaxwell(dat, firstq = "ITSS_1")
}
