% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMcBride.R
\name{scoreMcBride}
\alias{scoreMcBride}
\title{A function that scores McBride et al's Cognitive and Behavioral Outcomes of
Sexual Behavior Scale}
\usage{
scoreMcBride(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "CBOSBS_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates Cognitive and Behavioral subscale scores and total score
}
\examples{
dat = scoreMcBride(dat, firstq = "CBOSBS_1")
}
