% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMcDonaghMBISC.R
\name{scoreMcDonaghMBISC}
\alias{scoreMcDonaghMBISC}
\title{A function that scores McDonagh & Morrison's Male Body Image Self-Consciousness Scale}
\usage{
scoreMcDonaghMBISC(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "MBISCS_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates total score
}
\examples{
dat = scoreMcDonaghMBISC(dat, firstq = "MBISCS_1")
}
