% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMilhausenSESIMW.R
\name{scoreMilhausenSESIMW}
\alias{scoreMilhausenSESIMW}
\title{A function that scores Janssen et al.'s Sexual Excitation/Sexual Inhibition Inventory (short-form)}
\usage{
scoreMilhausenSESIMW(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "sesimw1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Subscale total scores
}
\examples{
dat = scoreMilhausenSESIMW(dat, firstq = "sesimw1")
}
