% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMosherGuilt.R
\name{scoreMosherGuilt}
\alias{scoreMosherGuilt}
\title{A function that scores the Revised Mosher's Guilt Inventory}
\usage{
scoreMosherGuilt(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "RMGI_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculate's sex-guilt, hostility-guilt and guity-conscience subscale scores.
}
\examples{
dat = scoreMosherGuilt(dat, firstq = "RMGI_1")
}
