% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMuehlenhardBeliefs.R
\name{scoreMuehlenhardBeliefs}
\alias{scoreMuehlenhardBeliefs}
\title{A function that scores Muehlenhard & Felts's Sexual Beliefs Scale}
\usage{
scoreMuehlenhardBeliefs(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "SBS_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates five subscale scores
}
\examples{
dat = scoreMuehlenhardBeliefs(dat, firstq = "SBS_1")
}
