% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorePascoal.R
\name{scorePascoal}
\alias{scorePascoal}
\title{A function that scores Pascoal et al.'s Beliefs About Sexual Function Scale}
\usage{
scorePascoal(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "BASEF_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates five subscale scores and total score
}
\examples{
dat = scorePascoal(dat, firstq = "BASEF_1")
}
