% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreRenaud.R
\name{scoreRenaud}
\alias{scoreRenaud}
\title{A function that scores Renaud & Byers Sexual Cognition Checklist}
\usage{
scoreRenaud(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "SCC_P_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates positive and negative total scores, as well as subscales for dominance and submission
}
\examples{
dat = scoreRenaud(dat, firstq = "SCC_P_1")
}
