% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreSalesWorry.R
\name{scoreSalesWorry}
\alias{scoreSalesWorry}
\title{A function that scores Sales et al.'s Worry About Sexual Outcome Scale}
\usage{
scoreSalesWorry(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "WASOS_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates pregnancy and STI/HIV subscales, as well as total worry
}
\examples{
dat = scoreSalesWorry(dat, firstq = "WASOS_1")
}
