% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreSchaefer.R
\name{scoreSchaefer}
\alias{scoreSchaefer}
\title{A function that scores Schaefer et al.'s Empathy for Children Scale}
\usage{
scoreSchaefer(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "Q2_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates scenario-, subscale-, and total-scores based on a data frame. Users specify first column
with relevant data, and function seeks out the rest automatically.
}
\examples{
dat = scoreSchaefer(dat, firstq = "Q2_1")
}
