% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreSchwartz.R
\name{scoreSchwartz}
\alias{scoreSchwartz}
\title{A function that scores Schwartz's First Coital Affective Reaction Scale}
\usage{
scoreSchwartz(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "FCARS")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates scale score for for those with experience or anticipating their first experience
}
\examples{
dat = scoreSmall(dat, firstq = "FCARS")
}
