% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreSeto.R
\name{scoreSeto}
\alias{scoreSeto}
\title{A function that scores Seto et al.'s Revised Screening Scale for Pedophilic Interests}
\usage{
scoreSeto(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
SPPI variable (e.g., "RSS_1")}
}
\value{
data frame with new "SPPI2" vector of total scores
}
\description{
Calculates total SPPI-2 scores based on a data frame. Users specify first column
with relevant data, and function seeks out the rest automatically.
}
\examples{
dat = scoreSeto(dat, firstq = "RSS_1")
}
