% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreSmall.R
\name{scoreSmall}
\alias{scoreSmall}
\title{A function that scores Small's Perceived Costs and Benefits Scale}
\usage{
scoreSmall(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "APCBSSI_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates costs and benefits subscale scores
}
\examples{
dat = scoreSmall(dat, firstq = "APCBSSI_1")
}
