% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreVanLankveld.R
\name{scoreVanLankveld}
\alias{scoreVanLankveld}
\title{A function that scores Napper's Negative Impact of Hookups Inventory}
\usage{
scoreVanLankveld(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "SSCS_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates scale score scores
}
\examples{
dat = scoreVanLankveld(dat, firstq = "SSCS_1")
}
