% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreVisser.R
\name{scoreVisser}
\alias{scoreVisser}
\title{A function that scores Visser et al.'s Male Enjoyment of Sexualization Scale}
\usage{
scoreVisser(dat, firstq)
}
\arguments{
\item{dat}{data frame}

\item{firstq}{input character for the name of the column for the first
variable (e.g., "MESS_1")}
}
\value{
data frame with new vectors of  scores
}
\description{
Calculates total score
}
\examples{
dat = scoreVisser(dat, firstq = "MESS_1")
}
