% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexscales-package.R
\docType{package}
\name{sexscales-package}
\alias{sexscales}
\alias{sexscales-package}
\title{sexscales: sexscales: Automatic scoring of scales in the Handbook of
    Sexuality-Related Measures (4th ed.)}
\description{
This package will take data collected with one
    or more of the pre-programmed Qualtrics surveys and automatically
    calculate and return a data frame with all (sub)scale scores.
}
\author{
\strong{Maintainer}: John Sakaluk \email{jksakaluk@gmail.com} (0000-0002-2515-9822)

}
\keyword{internal}
